Extract GameTrailer component from GameDetail
